// resultsdash/src/schemas.ts
const numberOfFunds = 252,
	numberOfAllocationSources = 4,
	numberOfOrganizationTypes = 4,
	numberOfBeneficiaryTypes = 23,
	numberOfSectors = 17;

const beneficiariesSchema = z.number().nullable();
// ...
export const bySectorObjSchema = z.object({
	PooledFundId: z.number().min(1).max(numberOfFunds),
	AllocationYear: z.number(),
	ReportApprovedDate: z.date(),
	AllocationtypeId: z.number(),
	AllocationSourceId: z.number().min(1).max(numberOfAllocationSources),
	ClusterId: z.number().min(1).max(numberOfSectors),
	ClusterBudget: z.number().min(0),
	TargetedMen: beneficiariesSchema,
	TargetedWomen: beneficiariesSchema,
	TargetedBoys: beneficiariesSchema,
	TargetedGirls: beneficiariesSchema,
	ReachedMen: beneficiariesSchema,
	ReachedWomen: beneficiariesSchema,
	ReachedBoys: beneficiariesSchema,
	ReachedGirls: beneficiariesSchema,
});
// ...
export const byTypeObjSchema = z.object({
	PooledFundId: z.number().min(1).max(numberOfFunds),
	AllocationYear: z.number(),
	ReportApprovedDate: z.date(),
	BeneficiaryTypeId: z.number().min(1).max(numberOfBeneficiaryTypes),
	AllocationtypeId: z.number(),
	AllocationSourceId: z.number().min(1).max(numberOfAllocationSources),
	TargetMen: beneficiariesSchema,
	TargetWomen: beneficiariesSchema,
	TargetBoys: beneficiariesSchema,
	TargetGirls: beneficiariesSchema,
	ReachedMen: beneficiariesSchema,
	ReachedWomen: beneficiariesSchema,
	ReachedBoys: beneficiariesSchema,
	ReachedGirls: beneficiariesSchema,
});

export const byOrganizationObjSchema = z.object({
	PooledFundId: z.number().min(1).max(numberOfFunds),
	AllocationYear: z.number(),
	ReportApprovedDate: z.date(),
	AllocationtypeId: z.number(),
	AllocationSourceId: z.number().min(1).max(numberOfAllocationSources),
	OrganizationType: z.number().min(1).max(numberOfOrganizationTypes),
	NumbofProjects: z.number(),
	TotalNumbPartners: z.number(),
	Budget: z.number().min(0),
	TargetedMen: beneficiariesSchema,
	TargetedWomen: beneficiariesSchema,
	TargetedBoys: beneficiariesSchema,
	TargetedGirls: beneficiariesSchema,
	ReachedMen: beneficiariesSchema,
	ReachedWomen: beneficiariesSchema,
	ReachedBoys: beneficiariesSchema,
	ReachedGirls: beneficiariesSchema,
});

export type ApprovedAllocationsObj = z.infer<
	typeof ApprovedAllocationsObjSchema
>;

export type BySectorObj = z.infer<typeof bySectorObjSchema>;

export type ByDisabilityObj = z.infer<typeof byDisabilityObjSchema>;

export type ByLocationObj = z.infer<typeof byLocationObjSchema>;

export type ByTypeObj = z.infer<typeof byTypeObjSchema>;

export type ByOrganizationObj = z.infer<typeof byOrganizationObjSchema>;